templates: unified manage
