+++ pkg/apiserver/apiserver.go
@@ -14,7 +14,6 @@
 package apiserver
 
 import (
-	"html/template"
 	"net/http"
 	"sync"
 
@@ -38,6 +37,7 @@ import (
 	"github.com/pingcap-incubator/tidb-dashboard/pkg/dbstore"
 	"github.com/pingcap-incubator/tidb-dashboard/pkg/keyvisual"
 	"github.com/pingcap-incubator/tidb-dashboard/pkg/tidb"
+	"github.com/pingcap-incubator/tidb-dashboard/templates"
 )
 
 var once sync.Once
@@ -62,8 +62,7 @@ func Handler(apiPrefix string, config *config.Config, services *Services) http.H
 	r.Use(gzip.Gzip(gzip.BestSpeed))
 	r.Use(utils.MWHandleErrors())
 
-	templates := template.New("api")
-	r.SetHTMLTemplate(templates)
+	templates.GinLoad(r)
 
 	endpoint := r.Group(apiPrefix)
 
@@ -77,7 +76,7 @@ func Handler(apiPrefix string, config *config.Config, services *Services) http.H
 	services.KeyVisual.Register(endpoint, auth)
 	logsearch.NewService(config, services.Store).Register(endpoint, auth)
 	statement.NewService(config, services.TiDBForwarder).Register(endpoint, auth)
-	diagnose.NewService(config, services.TiDBForwarder, services.Store).Register(endpoint, auth).RegisterTemplates(templates)
+	diagnose.NewService(config, services.TiDBForwarder, services.Store).Register(endpoint, auth)
 
 	return r
 }
